catalog improvements
create API serves, destruct catalog js file, improve container styles, add catalog section and title styles
