+++ src/partials/catalog.html
@@ -1,14 +1,11 @@
-<!-- 
-<select name="select" id="select">
-    <option value="A">A</option>
-    <option value="B">B</option>
-    <option value="C">C</option>
-</select>
-<button class="btn" type="button" value="a">a</button> -->
-
 <section class="catalog">
   <h2 class="catalog__title">Cocktails</h2>
   <div class="container">
-    <ul class="catalog__list"></ul>
+    <ul class="catalog__list">
+      <!-- cocktails content here -->
+    </ul>
+    <div class="paginator">
+      <!-- pagination buttons content here-->
+    </div>
   </div>
 </section>
