+++ src/sass/base/_container.scss
@@ -1,7 +1,7 @@
 @import '../utils/variables';
+@import '../utils/mixins';
 
 .container {
-    box-sizing: content-box;
     width: $mobile;
     max-width: 100%;
     padding-left: 20px;
@@ -9,13 +9,15 @@
     margin-left: auto;
     margin-right: auto;
 
-    @media screen and (min-width: $tablet) {
+    @include tablet {
         width: $tablet;
         padding-left: 32px;
         padding-right: 32px;
     }
 
-    @media screen and (min-width: $desktop) {
+    @include desktop {
         width: $desktop;
+        padding-left: 16px;
+        padding-right: 16px;
     }
 }
