+++ src/sass/layout/_catalog.scss
@@ -1,73 +1,111 @@
 @import '../utils/mixins';
 @import '../utils/variables';
 
+.catalog {
+  padding-bottom: 20px;
+
+  @include tablet {
+    padding-bottom: 60px;
+  }
+}
+
+.catalog__title {
+  padding-bottom: 60px;
+  font-family: 'Poppins';
+  font-style: normal;
+  font-weight: 600;
+  font-size: 22px;
+  line-height: 1.9;
+  text-align: center;
+  color: var(--text-primary-color);
+
+  @include tablet {
+    font-weight: 700;
+    font-size: 32px;
+    line-height: 1.3;
+  }
+
+  @include desktop {
+    font-size: 48px;
+    line-height: 1.5;
+  }
+}
+
 .catalog__list {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    gap: 30px;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: 30px;
 }
 
 .catalog__item {
-    border: 1px solid var(--text-accent-color);
-    border-radius: 4px;
+  border: 1px solid var(--text-accent-color);
+  border-radius: 4px;
 }
 
 .catalog__wrapper-img {
-    width: 280px;
-    height: 280px;
-    overflow: hidden;
-    border-radius: 4px 4px 0 0;
+  width: 280px;
+  height: 280px;
+  overflow: hidden;
+  border-radius: 4px 4px 0 0;
 
-    @include tablet {
-        width: 335px;
-        height: 395px;
-    }
-    
-    @include desktop {
-        width: 395px;
-        height: 395px;
-    }
+  @include tablet {
+    width: 335px;
+    height: 395px;
+  }
+
+  @include desktop {
+    width: 393px;
+    height: 395px;
+  }
 }
 
 .catalog__img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
 }
 
 .catalog__item-name {
-    font-style: normal;
-    font-weight: 900;
-    font-size: 26px;
-    line-height: 1.2;
-    text-align: center;
-    margin-bottom: 16px;
-    margin-top: 48px;
-    
-    @include tablet {
-        font-size: 28px;
-        margin-bottom: 25px;
-        margin-top: 20px;
-    }
+  margin-left: auto;
+  margin-right: auto;
+  font-style: normal;
+  font-weight: 900;
+  font-size: 26px;
+  line-height: 1.2;
+  text-align: center;
+  margin-bottom: 16px;
+  margin-top: 48px;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  max-width: 260px;
 
-    @include desktop {
-        font-size: 32px;
-        margin-bottom: 27px;
-    }
+  @include tablet {
+    font-size: 28px;
+    margin-bottom: 25px;
+    margin-top: 20px;
+    max-width: 315px;
+  }
+
+  @include desktop {
+    font-size: 32px;
+    margin-bottom: 27px;
+    max-width: 373px;
+  }
 }
 
 .catalog__item-actions {
-    display: flex;
-    justify-content: center;
-    gap: 16px;
-    padding-bottom: 24px;
+  display: flex;
+  justify-content: center;
+  gap: 16px;
+  padding-bottom: 24px;
 
-    @include tablet {
-        padding-bottom: 20px;
-    }
+  @include tablet {
+    padding-bottom: 20px;
+  }
 
-    @include desktop {
-        padding-bottom: 40px;
-    }
+  @include desktop {
+    padding-bottom: 40px;
+  }
 }
